requirements.txt
numpy
pillow
onnx
onnxruntime
